Revert check for maxFragmentCombinedOutputResources
WebGPU removed its maxFragmentCombinedOutputResources limit so now we
need to check it against the 8+4+8 instead of the default value of that
limit. This code is what we had before, except this also adds ImgTec to
the list of vendors that report this limit incorrectly. (The limit seems
to exist just for Adreno.)